+++ src/components/search-input/index.tsx
@@ -6,7 +6,7 @@ import { SearchIcon } from '@/components/svg/search-icon'
 import { Input } from '@/components/ui/input'
 import { IGDBGameSearchSuggestion } from '@/lib/igdb/types'
 import { cn } from '@/lib/utils'
-import { getIGDBImageUrl } from '@/services/igdb/imageService'
+import { getIGDBImageUrl } from '@/services/igdb/images'
 import Image from 'next/image'
 import Link from 'next/link'
 import { useSearchInput } from './use-search-input'
