+++ src/services/igdb/games/index.ts
@@ -1,2 +1,3 @@
 export * from './get-details'
+export * from './get-related-games'
 export * from './get-suggestions'
