--- src/services/igdb/index.ts
@@ -1,2 +1,0 @@
-export * from './games'
-export * from './imageService'
